make SparseJoinState visible in header file
